requirements.txt
paho-mqtt
python-astar
opencv-python
numpy
pygame